nvme: move nvme-mi commands into a plugin
There are too many top-level commands cluttering the space. Move the
nvme-mi commands into a plugin, e.g.,

  nvme nmve-mi-send -> nvme mi send
