+++ src/nvme-builtin.h
@@ -143,8 +143,10 @@ COMMAND_LIST(
 	ENTRY("show-topology", "Show the topology", show_topology_cmd)
 	ENTRY("io-mgmt-recv", "I/O Management Receive", io_mgmt_recv)
 	ENTRY("io-mgmt-send", "I/O Management Send", io_mgmt_send)
-	ENTRY("nvme-mi-recv", "Submit a NVMe-MI Receive command, return results", nmi_recv)
-	ENTRY("nvme-mi-send", "Submit a NVMe-MI Send command, return results", nmi_send)
+#ifdef CONFIG_DEPRECATED_CMDS
+	ENTRY_DEPRECATED("nvme-mi-recv", "Submit a NVMe-MI Receive command, return results (deprecated, use 'nvme nvme-mi recv')", nmi_recv)
+	ENTRY_DEPRECATED("nvme-mi-send", "Submit a NVMe-MI Send command, return results (deprecated, use 'nvme nvme-mi send')", nmi_send)
+#endif /* CONFIG_DEPRECATED_CMDS */
 );
 
 #endif
